Modified redirect urls and tests according to previous changes
