+++ web/src/main/java/com/epam/rft/atsy/web/controllers/NewApplicationController.java
@@ -56,6 +56,6 @@ public class NewApplicationController {
 
       applicationsService.saveApplication(applicationDTO, stateHistoryDTO);
     }
-    return "redirect:/secure/candidate/" + stateHistoryDTO.getCandidateId();
+    return "redirect:/secure/candidate/details/" + stateHistoryDTO.getCandidateId();
   }
 }
