+++ web/src/test/java/com/epam/rft/atsy/web/controllers/CandidateCreationControllerTest.java
@@ -23,15 +23,17 @@ public class CandidateCreationControllerTest extends AbstractControllerTest {
 
   private static final String ERROR_VIEW_NAME = "error";
 
-  private static final String CANDIDATE_CREATION_REQUEST_URL = "/secure/candidate";
+  private static final String CANDIDATE_CREATION_REQUEST_URL = "/secure/candidate/details";
 
   private static final Long CANDIDATE_ID = 1L;
 
+  private static final Long NON_EXISTENT_CANDIDATE_ID = -1L;
+
   private static final String EXISTING_CANDIDATE_REQUEST_URL =
-      "/secure/candidate/" + CANDIDATE_ID.toString();
+      "/secure/candidate/details/" + CANDIDATE_ID.toString();
 
   private static final String NON_EXISTENT_CANDIDATE_REQUEST_URL =
-      "/secure/candidate/" + CANDIDATE_ID.toString();
+      "/secure/candidate/details/" + NON_EXISTENT_CANDIDATE_ID.toString();
 
   private static final String CANDIDATE_OBJECT_KEY = "candidate";
 
@@ -86,13 +88,14 @@ public class CandidateCreationControllerTest extends AbstractControllerTest {
   @Test
   public void loadCandidateShouldRenderErrorViewWhenThereIsNoCandidateWithIdInPathParam()
       throws Exception {
-    given(candidateService.getCandidate(CANDIDATE_ID)).willThrow(IllegalArgumentException.class);
+    given(candidateService.getCandidate(NON_EXISTENT_CANDIDATE_ID))
+        .willThrow(IllegalArgumentException.class);
 
     mockMvc.perform(get(NON_EXISTENT_CANDIDATE_REQUEST_URL))
         .andExpect(status().isInternalServerError())
         .andExpect(view().name(ERROR_VIEW_NAME))
         .andExpect(forwardedUrl(VIEW_PREFIX + ERROR_VIEW_NAME + VIEW_SUFFIX));
 
-    then(candidateService).should().getCandidate(CANDIDATE_ID);
+    then(candidateService).should().getCandidate(NON_EXISTENT_CANDIDATE_ID);
   }
 }
